All moves and captures working correctly
Accept entpassant works correctly and peices move as expected.
All input conflicts resolved.
