+++ Assets/Scenes/GameScene2.unity
@@ -407,7 +407,7 @@ SpriteRenderer:
   m_WasSpriteAssigned: 1
   m_MaskInteraction: 0
   m_SpriteSortPoint: 0
---- !u!1 &468286562
+--- !u!1 &465824444
 GameObject:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -415,9 +415,8 @@ GameObject:
   m_PrefabAsset: {fileID: 0}
   serializedVersion: 6
   m_Component:
-  - component: {fileID: 468286564}
-  - component: {fileID: 468286563}
-  - component: {fileID: 468286565}
+  - component: {fileID: 465824447}
+  - component: {fileID: 465824446}
   m_Layer: 0
   m_Name: Pieces
   m_TagString: Untagged
@@ -425,13 +424,13 @@ GameObject:
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!114 &468286563
+--- !u!114 &465824446
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 468286562}
+  m_GameObject: {fileID: 465824444}
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 415cc07ab1220fa47966e20f90650170, type: 3}
@@ -449,35 +448,22 @@ MonoBehaviour:
   blackBishop: {fileID: 138526475489352158, guid: 7ecabf2fb704922448c79a6c9af1cdec, type: 3}
   blackQueen: {fileID: 3515128362262936837, guid: 673171e4a5666604da2cce8454819825, type: 3}
   blackKing: {fileID: 4693576580125088435, guid: ce586da612187ac4aaff53d6bd5608ed, type: 3}
-  positionZ: -1
-  occupiedPositionsList: []
---- !u!4 &468286564
+  positionZ: 0
+--- !u!4 &465824447
 Transform:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 468286562}
+  m_GameObject: {fileID: 465824444}
   serializedVersion: 2
-  m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 0, z: 0}
-  m_LocalScale: {x: 0.5, y: 0.5, z: 1}
+  m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 1695861753}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &468286565
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 468286562}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 30dff641abbc2ec48862a448d948fbaa, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!1 &889691940
 GameObject:
   m_ObjectHideFlags: 0
@@ -879,7 +865,6 @@ GameObject:
   m_Component:
   - component: {fileID: 1113360544}
   - component: {fileID: 1113360545}
-  - component: {fileID: 1113360546}
   m_Layer: 0
   m_Name: BoardTiles
   m_TagString: Untagged
@@ -914,22 +899,9 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 872f6fb7d1e6fea42ab9712cf52a1b6c, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  piecePlacement: {fileID: 468286563}
   tilePrefab: {fileID: 5762455892759881398, guid: e5326f5b3738fd74cad37f0da87e75c5, type: 3}
   blackTileColor: {r: 0.18867922, g: 0, b: 0, a: 1}
   whiteTileColor: {r: 0.8584906, g: 0.8584906, b: 0.8584906, a: 1}
---- !u!114 &1113360546
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1113360543}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 27a3a21572681eb4cb54a22e898db5be, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!1 &1412039928
 GameObject:
   m_ObjectHideFlags: 0
@@ -1152,6 +1124,8 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 1695861753}
+  - component: {fileID: 1695861755}
+  - component: {fileID: 1695861756}
   m_Layer: 0
   m_Name: ChessBoard
   m_TagString: Untagged
@@ -1174,9 +1148,40 @@ Transform:
   m_Children:
   - {fileID: 416304019}
   - {fileID: 1113360544}
-  - {fileID: 468286564}
+  - {fileID: 465824447}
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &1695861755
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1695861751}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: ce872bbdf110f6245b3fdd9271f22814, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  occupiedPositionsList: []
+--- !u!114 &1695861756
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1695861751}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 30dff641abbc2ec48862a448d948fbaa, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  boardLayer:
+    serializedVersion: 2
+    m_Bits: 8
+  pieceLayer:
+    serializedVersion: 2
+    m_Bits: 64
 --- !u!1 &1784941429
 GameObject:
   m_ObjectHideFlags: 0
